refactor: 이미지를 가져오는 형식 변경
