+++ client/src/lib/common/avatar/index.ts
@@ -1,107 +1,35 @@
-import Boy0 from '@public/image/Boy-0.svg';
-import Boy1 from '@public/image/Boy-1.svg';
-import Boy2 from '@public/image/Boy-2.svg';
-import Boy3 from '@public/image/Boy-3.svg';
-import Boy4 from '@public/image/Boy-4.svg';
-import Boy5 from '@public/image/Boy-5.svg';
-import Boy6 from '@public/image/Boy-6.svg';
-import Boy7 from '@public/image/Boy-7.svg';
-import Boy8 from '@public/image/Boy-8.svg';
-import Boy9 from '@public/image/Boy-9.svg';
-import Boy10 from '@public/image/Boy-10.svg';
-import Boy11 from '@public/image/Boy-11.svg';
-import Boy12 from '@public/image/Boy-12.svg';
-import Boy13 from '@public/image/Boy-13.svg';
-import Boy14 from '@public/image/Boy-14.svg';
-import Boy15 from '@public/image/Boy-15.svg';
-import Boy16 from '@public/image/Boy-16.svg';
-import Girl0 from '@public/image/Girl-0.svg';
-import Girl1 from '@public/image/Girl-1.svg';
-import Girl2 from '@public/image/Girl-2.svg';
-import Girl3 from '@public/image/Girl-3.svg';
-import Girl4 from '@public/image/Girl-4.svg';
-import Girl5 from '@public/image/Girl-5.svg';
-import Girl6 from '@public/image/Girl-6.svg';
-import Girl7 from '@public/image/Girl-7.svg';
-import Girl8 from '@public/image/Girl-8.svg';
-import Girl9 from '@public/image/Girl-9.svg';
-import Girl10 from '@public/image/Girl-10.svg';
-import Girl11 from '@public/image/Girl-11.svg';
-import Girl12 from '@public/image/Girl-12.svg';
-import Girl13 from '@public/image/Girl-13.svg';
-import Girl14 from '@public/image/Girl-14.svg';
-import Girl15 from '@public/image/Girl-15.svg';
-import Girl16 from '@public/image/Girl-16.svg';
+export { default as Boy0 } from '@public/image/Boy-0.svg';
+export { default as Boy1 } from '@public/image/Boy-1.svg';
+export { default as Boy2 } from '@public/image/Boy-2.svg';
+export { default as Boy3 } from '@public/image/Boy-3.svg';
+export { default as Boy4 } from '@public/image/Boy-4.svg';
+export { default as Boy5 } from '@public/image/Boy-5.svg';
+export { default as Boy6 } from '@public/image/Boy-6.svg';
+export { default as Boy7 } from '@public/image/Boy-7.svg';
+export { default as Boy8 } from '@public/image/Boy-8.svg';
+export { default as Boy9 } from '@public/image/Boy-9.svg';
+export { default as Boy10 } from '@public/image/Boy-10.svg';
+export { default as Boy11 } from '@public/image/Boy-11.svg';
+export { default as Boy12 } from '@public/image/Boy-12.svg';
+export { default as Boy13 } from '@public/image/Boy-13.svg';
+export { default as Boy14 } from '@public/image/Boy-14.svg';
+export { default as Boy15 } from '@public/image/Boy-15.svg';
+export { default as Boy16 } from '@public/image/Boy-16.svg';
 
-export type ImageType =
-  | 'Boy0'
-  | 'Boy1'
-  | 'Boy2'
-  | 'Boy3'
-  | 'Boy4'
-  | 'Boy5'
-  | 'Boy6'
-  | 'Boy7'
-  | 'Boy8'
-  | 'Boy9'
-  | 'Boy10'
-  | 'Boy11'
-  | 'Boy12'
-  | 'Boy13'
-  | 'Boy14'
-  | 'Boy15'
-  | 'Boy16'
-  | 'Girl0'
-  | 'Girl1'
-  | 'Girl2'
-  | 'Girl3'
-  | 'Girl4'
-  | 'Girl5'
-  | 'Girl6'
-  | 'Girl7'
-  | 'Girl8'
-  | 'Girl9'
-  | 'Girl10'
-  | 'Girl11'
-  | 'Girl12'
-  | 'Girl13'
-  | 'Girl14'
-  | 'Girl15'
-  | 'Girl16';
-
-export default {
-  Boy0,
-  Boy1,
-  Boy2,
-  Boy3,
-  Boy4,
-  Boy5,
-  Boy6,
-  Boy7,
-  Boy8,
-  Boy9,
-  Boy10,
-  Boy11,
-  Boy12,
-  Boy13,
-  Boy14,
-  Boy15,
-  Boy16,
-  Girl0,
-  Girl1,
-  Girl2,
-  Girl3,
-  Girl4,
-  Girl5,
-  Girl6,
-  Girl7,
-  Girl8,
-  Girl9,
-  Girl10,
-  Girl11,
-  Girl12,
-  Girl13,
-  Girl14,
-  Girl15,
-  Girl16,
-};
+export { default as Girl0 } from '@public/image/Girl-0.svg';
+export { default as Girl1 } from '@public/image/Girl-1.svg';
+export { default as Girl2 } from '@public/image/Girl-2.svg';
+export { default as Girl3 } from '@public/image/Girl-3.svg';
+export { default as Girl4 } from '@public/image/Girl-4.svg';
+export { default as Girl5 } from '@public/image/Girl-5.svg';
+export { default as Girl6 } from '@public/image/Girl-6.svg';
+export { default as Girl7 } from '@public/image/Girl-7.svg';
+export { default as Girl8 } from '@public/image/Girl-8.svg';
+export { default as Girl9 } from '@public/image/Girl-9.svg';
+export { default as Girl10 } from '@public/image/Girl-10.svg';
+export { default as Girl11 } from '@public/image/Girl-11.svg';
+export { default as Girl12 } from '@public/image/Girl-12.svg';
+export { default as Girl13 } from '@public/image/Girl-13.svg';
+export { default as Girl14 } from '@public/image/Girl-14.svg';
+export { default as Girl15 } from '@public/image/Girl-15.svg';
+export { default as Girl16 } from '@public/image/Girl-16.svg';
