+++ client/src/types/image.d.ts
@@ -1,0 +1,35 @@
+export type ImageType =
+  | 'Boy0'
+  | 'Boy1'
+  | 'Boy2'
+  | 'Boy3'
+  | 'Boy4'
+  | 'Boy5'
+  | 'Boy6'
+  | 'Boy7'
+  | 'Boy8'
+  | 'Boy9'
+  | 'Boy10'
+  | 'Boy11'
+  | 'Boy12'
+  | 'Boy13'
+  | 'Boy14'
+  | 'Boy15'
+  | 'Boy16'
+  | 'Girl0'
+  | 'Girl1'
+  | 'Girl2'
+  | 'Girl3'
+  | 'Girl4'
+  | 'Girl5'
+  | 'Girl6'
+  | 'Girl7'
+  | 'Girl8'
+  | 'Girl9'
+  | 'Girl10'
+  | 'Girl11'
+  | 'Girl12'
+  | 'Girl13'
+  | 'Girl14'
+  | 'Girl15'
+  | 'Girl16';
